consolidated v2.0 spec support, added v2.0 spec links
